+++ systemh-housesim/src/edu/hawaii/systemh/housesimulator/lighting/living/LightingLivingPutResource.java
@@ -54,7 +54,7 @@ public class LightingLivingPutResource extends ServerResource {
     
     // Call mutator corresponding to room.
     if (setLevel.equalsIgnoreCase(putCommand) && "SET_LIGHTING_LEVEL".equalsIgnoreCase(command)) {
-      LightingData.setLivingLevel(Long.parseLong(arg));
+      LightingData.setLivingLevel(Integer.parseInt(arg));
     }
     // Call lighting mutator corresponding to room.
     else if (setColor.equalsIgnoreCase(putCommand) &&
